requirements.txt
pandas
requests
python-dotenv
minio
psycopg2-binary
mysql-connector-python
kaggle
pytest
apache-airflow
boto3
python-dotenv
